리덕스 store, reducer 폴더 구조 변경
 - 기존 Reducer 의 경우 Ducks 구조로 이루어진 Module(reducer + action) 로 볼 수 있음
 - store, reducer 는 module 로 통합
 - 기타 로직 정리
