+++ src/pages/index.tsx
@@ -1,8 +1,9 @@
 import * as React from 'react';
+import { NextPage } from 'next';
 
 import GnbLayout from 'components/templates/GnbLayout';
 
-const Home: React.FunctionComponent = () => {
+const Home: NextPage = () => {
     return (
         <GnbLayout title={'Home'}>
             <h1>Hello, Next Ts</h1>
