--- src/store/actions.ts
@@ -1,1 +1,0 @@
-export { PostsAction } from 'store/reducers/PostsReducer';
