+++ src/types/TStoreState.ts
@@ -1,0 +1,5 @@
+import rootReducer from 'modules/reducer';
+
+type TStoreState = ReturnType<typeof rootReducer>;
+
+export default TStoreState;
